fix: don't throw exception in BiomeManager#checkBiomeChangeEvent
The `BiomeManager` was very picky about inconsistencies of assigned
biomes, i.e., it would throw an exception if there is biome information
for one position but not for another.

As this mainly occurs when the player enters an not-yet-loaded chunk
throwing an exception is a bit overly dramatic. We can live with logging
a WARNING here (to keep being informed about this).

We may later reduce the logging level or even get rid of it completely.